+++ src/mat3.rs
@@ -10,7 +10,7 @@ use crate::vec::{Vector, VectorTrait};
 /// Stored As Row Major
 pub struct Mat3
 {
-    contents : [f32; 3 * 3]
+    contents : [f64; 3 * 3]
 }
 
 impl Mat3
@@ -62,14 +62,14 @@ impl Mat3
     }
     
     /// Change element at x, y (0 indexed)
-    pub fn change(&mut self, x : usize, y : usize, val : f32)
+    pub fn change(&mut self, x : usize, y : usize, val : f64)
     {
         let index = Self::index(x, y);
         self.contents[index] = val;
     }
 
     /// Get element at x, y (0 indexed)
-    pub fn get(&self, x : usize, y : usize) -> f32
+    pub fn get(&self, x : usize, y : usize) -> f64
     {
         let index = Self::index(x, y);
         self.contents[index]
@@ -115,7 +115,7 @@ impl Mat3
     }
 
     /// From Array (row major)
-    pub const fn from_array(mat : [f32; 3 * 3]) -> Mat3
+    pub const fn from_array(mat : [f64; 3 * 3]) -> Mat3
     {
         Mat3 { contents: mat }
     }
@@ -133,7 +133,7 @@ impl Mat3
     /// Create a rotation Matrix
     /// 
     /// http://www.songho.ca/opengl/gl_matrix.html
-    pub fn rotation_matrix(angle : f32, axis : Vector) -> Mat3
+    pub fn rotation_matrix(angle : f64, axis : Vector) -> Mat3
     {
         let s = angle.sin();
         let c = angle.cos();
@@ -151,7 +151,7 @@ impl Mat3
     }
 
     /// Rotate current matrix
-    pub fn rotate(&mut self, angle : f32, axis : Vector)
+    pub fn rotate(&mut self, angle : f64, axis : Vector)
     {
         *self = *self * Mat3::rotation_matrix(angle, axis);
     }
@@ -173,7 +173,7 @@ impl Mat3
     }
 
     /// get contents in array form
-    pub fn get_contents(&self) -> [f32; 3 * 3]
+    pub fn get_contents(&self) -> [f64; 3 * 3]
     {
         self.contents
     }
